Follow up fixes for WL#10634
The rework of the admin tabs left them in a bad state on macOS, due to some pecularities there. This patch changes some of the GUI construction code to take these pecularities into account when running on macOS. Additionally, I had to fix the tab drawing for secondary tabs on macOS as the old code did not properly account for title sizes and available space.

+++ library/grt/src/grtpp_module_python.cpp
@@ -246,8 +246,8 @@ static std::string formatStringList(PyObject *list) {
     std::string result;
     PyObject *item;
 
-    int count = PyList_Size(list);
-    for (int index = 0; index < count; ++index) {
+    ssize_t count = PyList_Size(list);
+    for (ssize_t index = 0; index < count; ++index) {
         item = PyList_GetItem(list, index);
         result += PyString_AsString(item);
     }
